refacotr: Database<T>에 공통 메서드 정의, UserDao 인터페이스에 추가적인 메서드 정의 후 UserDaoImpl에서 구현하여 사용하도록 변경함
